Move `slugify` from `utils` to `plot`
This function is used only in this module.
