+++ analysis/plot.py
@@ -2,7 +2,9 @@
 """
 import collections
 import pathlib
+import re
 import textwrap
+import unicodedata
 
 import click
 import pandas
@@ -65,7 +67,7 @@ def main(from_date, from_offset, d_out):
 
     figs_cols = plot(by_day, by_week, get_plot_title(from_date, from_offset))
     for fig, col in figs_cols:
-        f_stem = utils.slugify(col)
+        f_stem = slugify(col)
         fig.savefig(d_out / f"{f_stem}.png")
 
 
@@ -134,6 +136,21 @@ def plot(by_day, by_week, plot_title):
         ax.legend(loc="upper right")
 
         yield fig, col
+
+
+def slugify(s):
+    # Based on Django's slugify. For more information, see:
+    # https://github.com/django/django/blob/4.1.7/django/utils/text.py#L399-L417
+
+    # convert to ASCII
+    s = unicodedata.normalize("NFKD", s).encode("ascii", "ignore").decode("ascii")
+    # remove characters that are not word, white space, or dash
+    s = re.sub(r"[^\w\s-]", "", s)
+    # replace one or more dash or one or more white space with one dash
+    s = re.sub(r"[-\s]+", "-", s)
+    # remove leading and trailing dashes and underscores
+    s = s.strip("-_")
+    return s.lower()
 
 
 if __name__ == "__main__":
