--- tests/test_utils.py
@@ -1,8 +1,0 @@
-import pytest
-
-from analysis import utils
-
-
-@pytest.mark.parametrize("string,slug", [("Ça va?", "ca-va"), ("_so--so_", "so-so")])
-def test_slugify(string, slug):
-    assert utils.slugify(string) == slug
